프로젝트 main에서 pull

+++ insider/src/main/java/com/kh/insider/dto/ReportManagementDto.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class ReportManagementDto {
 	private long reportMemberNo;
-	private int reportTableNo;
+	private long reportTableNo;
 	private String reportTable;
 	private Integer reportResult;
 	private int count;
